Split source DB tag catalog
